requirements.txt
numpy>=1.24
opencv-python-headless>=4.8
Pillow>=10.0
gymnasium>=1.0
stable-baselines3[extra]>=2.0
torch>=2.0
psutil>=5.9
